Update .gitignore, add langgraph.json, and refactor markdown normalization (#454)
